requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pandas==2.2.1
numpy==1.26.3
openpyxl==3.1.2
fpdf==1.7.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
